+++ database/postgres/config.go
@@ -9,96 +9,102 @@ import (
 )
 
 func init() {
+	viper.SetDefault(dbUserViperKey, dbUserDefault)
 	_ = viper.BindEnv(dbUserViperKey, dbUserEnv)
+	viper.SetDefault(dbPasswordViperKey, dbPasswordDefault)
 	_ = viper.BindEnv(dbPasswordViperKey, dbPasswordEnv)
+	viper.SetDefault(dbDatabaseViperKey, dbDatabaseDefault)
 	_ = viper.BindEnv(dbDatabaseViperKey, dbDatabaseEnv)
+	viper.SetDefault(dbHostViperKey, dbHostDefault)
 	_ = viper.BindEnv(dbHostViperKey, dbHostEnv)
+	viper.SetDefault(dbPortViperKey, dbPortDefault)
 	_ = viper.BindEnv(dbPortViperKey, dbPortEnv)
+	viper.SetDefault(dbPoolSizeViperKey, dbPoolSizeDefault)
 	_ = viper.BindEnv(dbPoolSizeViperKey, dbPoolSizeEnv)
 }
 
 // PGFlags register flags for Postgres database
 func PGFlags(f *pflag.FlagSet) {
-	DBUser(f, "postgres")
-	DBPassword(f, "postgres")
-	DBDatabase(f, "postgres")
-	DBHost(f, "127.0.0.1")
-	DBPort(f, 5432)
+	DBUser(f)
+	DBPassword(f)
+	DBDatabase(f)
+	DBHost(f)
+	DBPort(f)
 	DBPoolSize(f)
 }
 
 var (
 	dbUserFlag     = "db-user"
 	dbUserViperKey = "db.user"
+	dbUserDefault  = "postgres"
 	dbUserEnv      = "DB_USER"
 )
 
 // DBUser register flag for db user
-func DBUser(f *pflag.FlagSet, defaultUser string) {
+func DBUser(f *pflag.FlagSet) {
 	desc := fmt.Sprintf(`Database User.
 Environment variable: %q`, dbUserEnv)
-	f.String(dbUserFlag, defaultUser, desc)
+	f.String(dbUserFlag, dbUserDefault, desc)
 	_ = viper.BindPFlag(dbUserViperKey, f.Lookup(dbUserFlag))
-	viper.SetDefault(dbUserViperKey, defaultUser)
 }
 
 var (
 	dbPasswordFlag     = "db-password"
 	dbPasswordViperKey = "db.password"
+	dbPasswordDefault  = "postgres"
 	dbPasswordEnv      = "DB_PASSWORD"
 )
 
 // DBPassword register flag for db password
-func DBPassword(f *pflag.FlagSet, defaultPassword string) {
+func DBPassword(f *pflag.FlagSet) {
 	desc := fmt.Sprintf(`Database User password
 Environment variable: %q`, dbPasswordEnv)
-	f.String(dbPasswordFlag, defaultPassword, desc)
-	viper.SetDefault(dbPasswordViperKey, defaultPassword)
+	f.String(dbPasswordFlag, dbPasswordDefault, desc)
 	_ = viper.BindPFlag(dbPasswordViperKey, f.Lookup(dbPasswordFlag))
 }
 
 var (
 	dbDatabaseFlag     = "db-database"
 	dbDatabaseViperKey = "db.database"
+	dbDatabaseDefault  = "postgres"
 	dbDatabaseEnv      = "DB_DATABASE"
 )
 
 // DBDatabase register flag for db database name
-func DBDatabase(f *pflag.FlagSet, defaultDatabase string) {
+func DBDatabase(f *pflag.FlagSet) {
 	desc := fmt.Sprintf(`Target Database name
 Environment variable: %q`, dbDatabaseEnv)
-	f.String(dbDatabaseFlag, defaultDatabase, desc)
-	viper.SetDefault(dbDatabaseViperKey, defaultDatabase)
+	f.String(dbDatabaseFlag, dbDatabaseDefault, desc)
 	_ = viper.BindPFlag(dbDatabaseViperKey, f.Lookup(dbDatabaseFlag))
 }
 
 var (
 	dbHostFlag     = "db-host"
 	dbHostViperKey = "db.host"
+	dbHostDefault  = "127.0.0.1"
 	dbHostEnv      = "DB_HOST"
 )
 
 // DBHost register flag for database host
-func DBHost(f *pflag.FlagSet, defaultHost string) {
+func DBHost(f *pflag.FlagSet) {
 	desc := fmt.Sprintf(`Database host
 Environment variable: %q`, dbHostEnv)
-	f.String(dbHostFlag, defaultHost, desc)
-	viper.SetDefault(dbHostViperKey, defaultHost)
+	f.String(dbHostFlag, dbHostDefault, desc)
 	_ = viper.BindPFlag(dbHostViperKey, f.Lookup(dbHostFlag))
 }
 
 var (
 	dbPortFlag     = "db-port"
 	dbPortViperKey = "db.port"
+	dbPortDefault  = 5432
 	dbPortEnv      = "DB_PORT"
 )
 
 // DBPort register flag for database port
-func DBPort(f *pflag.FlagSet, defaultPort int) {
+func DBPort(f *pflag.FlagSet) {
 	desc := fmt.Sprintf(`Database port
 Environment variable: %q`, dbPortEnv)
-	f.Int(dbPortFlag, defaultPort, desc)
-	viper.SetDefault(dbPortViperKey, defaultPort)
+	f.Int(dbPortFlag, dbPortDefault, desc)
 	_ = viper.BindPFlag(dbPortViperKey, f.Lookup(dbPortFlag))
 }
 
@@ -114,7 +120,6 @@ func DBPoolSize(f *pflag.FlagSet) {
 	desc := fmt.Sprintf(`Maximum number of connections on database
 Environment variable: %q`, dbPoolSizeEnv)
 	f.Int(dbPoolSizeFlag, dbPoolSizeDefault, desc)
-	viper.SetDefault(dbPoolSizeViperKey, dbPoolSizeDefault)
 	_ = viper.BindPFlag(dbPoolSizeViperKey, f.Lookup(dbPoolSizeFlag))
 }